+++ CHANGES.md
@@ -2,8 +2,8 @@
 
 ## 0.7.0
 
-* BREAKING - changed arguments for `ed.updatePlaceholder` to `id, {status, progress, errored}`
-  * Placholder block `metadata.errored` will show in red and have an "×" button to remove
+* BREAKING - changed arguments for `ed.updatePlaceholder` to `id, {status, progress, failed}`
+  * Placholder block `metadata.failed` will show in red and have an "×" button to remove
 
 ## 0.6.2 - 2016-03-20